+++ DreamWorld_Server/include/Utility/Thread/ThWorkerJob.h
@@ -36,7 +36,7 @@ class ThWorkerJob : public OVERLAPPED {
 
   ThWorkerJob(WokerPtr&& workItem, WORKER_TYPE type);
 
-  void operator()(const DWORD ioByte, const uint64_t errorCode);
+  void operator()(const DWORD ioByte, const DWORD errorCode);
 
   WORKER_TYPE GetType() const {
     return m_type;
